Refactor #56 : 전체 구조 변경 및 오류 수정
- LoginPage 삭제, index.jsx로 이전
- 일부 동작하지 않는 함수 삭제
- (fix) 회원등록시 enter입력시 결과창 반복출력 수정
